siphon-client/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "siphon_client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "headwater-client",
    "httpx",
    "pytest-watch>=4.2.0",
    "python-dateutil>=2.8.2",
    "rich>=14.2.0",
    "siphon-api",
    "siphon-server",
]

[tool.hatch.build.targets.wheel]
packages = ["src/siphon_client"]
sources = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src/siphon_client/py.typed"]

[tool.uv.sources]
siphon_api = { path = "../siphon-api", editable = true }
siphon_server = { path = "../siphon-server", editable = true }
headwater_client = { path = "../../headwater/headwater-client", editable = true }

[project.scripts]
siphon = "siphon_client.cli.siphon_cli:main"
